refactoring: moving client Details around
